ar6000: fix Huawei butchering
In the actual release, the ifdef around "unsigned int softmac[6]"
was commented out, which caused an unused variable warning. It wasn't
merged as a result, though the debug messages they added (in code they
didn't seem to be using) were.

Clearly, this code is meant to be used as softmac sets incorrect addresses
based on the contents of the softmac file without it, though ath6kl_fetch_mac_file
should be used instead. This should be as Huawei intended and how it works
in stock firmware.